+++ examples/AdWords/v201802/AdvancedOperations/AddDynamicPageFeed.php
@@ -272,9 +272,8 @@ class AddDynamicPageFeed
         $urlAttributeValue = new FeedItemAttributeValue();
         $urlAttributeValue->setFeedAttributeId($feedDetails->urlAttributeId);
 
-        // Optional: Add the {feeditem} valuetrack parameter to track which page
-        // feed items lead to each click.
-        $url = $url . '?id={feeditem}';
+        // See https://support.google.com/adwords/answer/7166527 for page feed
+        // URL recommendations and rules.
         $urlAttributeValue->setStringValues([$url]);
 
         $labelAttributeValue = new FeedItemAttributeValue();